notification list & create feed bloc 적용

+++ lib/providers/notification_list_bloc.dart
@@ -25,6 +25,7 @@ class NotificationListBloc
   }
 
   Future<void> _clearData(ClearNotificationListEvent event, emit) async {
+    if (state.list.isEmpty) return;
     emit(state.copyWith(status: NotificationListStatus.success, list: []));
     await AlarmService.deleteAllAlarmList(userId: userId);
   }
